Fix All Failing Shinytest Diffs and Merge All gha- Branches

Function to walk the user through merging all gha- branches. This function will NOT push to the repository. Users must git push themselves.

fix_all_gha_branches(
  dir = "apps",
  sha = git_sha(dir),
  ...,
  save_results = NULL,
  ask_apps = FALSE,
  ask_branches = TRUE,
  ask_if_not_main = TRUE,
  accept_all = FALSE,
  repo_dir = file.path(dir, "..")
)

Arguments

dir

Root app folder path

sha

git sha of base branch to look for

...

Extra arguments passed to shinytest::viewTestDiff

save_results

Logical which commits and merges all branches after viewing test results

ask_apps, ask_branches

Logical which allows for particular apps branches to be inspected

ask_if_not_main

Logical which will check if main is the base branch

accept_all

Logical which will accept all changes for the apps selected when ask_apps = TRUE. This parameter is only recommended for use when the shinytest differences are known and can not be displayed using the shiny application.

repo_dir

Root repo folder path

Details

This function will NOT fix shinyjster failures.

Outline of steps performed:

Validate there are no git changes or untracked files in the current base branch

Validate there are no unexpected shinytest folders

If isTRUE(ask), ask for which branches to work with

Find all apps to update

For each failing shinytest app, accept / reject shinytest changes in each gha- branch. Add an individualized commit message including the app and environment flavor information

For each gha- branch, validate there are no unexpected shinytest folders. (Should be no unexpected folders after accepting / rejecting apps.)

For each gha- branch, merge into the base branch.

For each gha- branch, delete the locally checked out gha- branch. (Cleans up the local repo.)

Tell the user to call git push.