Test apps using shiny::runTests()

Test apps using shiny::runTests()

test_runtests(
  dir = "apps",
  apps = apps_runtests(dir, filter = filter),
  filter = NULL,
  assert = TRUE,
  timeout = as.difftime(10, units = "mins"),
  retries = 2,
  update_pkgs = TRUE,
  update_app_pkgs = TRUE
)

test_shinytest(
  dir = "apps",
  apps = apps_shinytest(dir),
  assert = TRUE,
  retries = 3
)

test_shinyjster(
  dir = "apps",
  apps = apps_shinyjster(dir),
  assert = TRUE,
  retries = 3
)

test_testthat(
  dir = "apps",
  apps = apps_testthat(dir),
  assert = TRUE,
  retries = 3
)

Arguments

dir
base folder to look for applications

apps
applications within dir to run

filter
filter to use on test file within application

assert
logical value which will determine if assert_runtests() will be called on the result

timeout
Length of time allowed for an application's full test suit can run before determining it is a failure

retries
number of attempts to retry before declaring the test a failure

update_pkgs
Logical value which will try to install all required shiny packages used for testing

update_app_pkgs
Logical value which will try to install all required app packages used for testing

Functions

test_runtests: Generic method to call all testing files

test_shinytest: Only execute shinytest.R test files

test_shinyjster: Only execute shinyjster.R test files

test_testthat: Only execute testtat.R test files