Test deployed apps

Automatically runs the next app in a fresh callr::r_bg session. To stop, close the shiny application window.

test_in_connect(
  dir = "apps",
  urls = connect_urls_cache(dir = dir, apps = apps, account = "barret", server =
    "beta.rstudioconnect.com"),
  server = attr(urls, "server"),
  account = attr(urls, "account"),
  apps = apps_manual(dir),
  app = apps[1],
  port = 8080,
  host = "127.0.0.1"
)

connect_urls_cache_save(
  dir = "apps",
  apps = apps_deploy(dir),
  server = "beta.rstudioconnect.com",
  account = "barret",
  save_file = connect_urls_cache_file(dir = dir, account = account, server = server),
  urls = connect_urls(dir = dir, apps = apps, account = account, server = server)
)

connect_urls_cache_file(dir, account, server)

connect_urls_cache(
  dir = "apps",
  apps = apps_deploy(dir),
  account = "barret",
  server = "beta.rstudioconnect.com"
)

test_in_shinyappsio(
  dir = "apps",
  apps = apps_manual(dir),
  app = apps[1],
  port = 8080,
  host = "127.0.0.1",
  account = "testing-apps",
  server = "shinyapps.io"
)

Arguments

dir
base folder to look for applications

urls
urls to be supplied by connect_urls output

server, account
Parameters that could be supplied to [rsconnect::deployApp]

apps
applications within dir to run

app
app number or name to start with. If numeric, it will match the leading number in the testing application

port
The TCP port that the application should listen on. If the port is not specified, and the shiny.port option is set (with options(shiny.port = XX)), then that port will be used. Otherwise, use a random port.

host
The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not. See Details.

save_file
location to save the file

Functions

test_in_connect: Test connect applications given the server and account

connect_urls_cache_save: Save connect urls as they can not be retrieved by everyone at run time

connect_urls_cache_file: Get cached connect url information or retrieve them from server

connect_urls_cache: Get cached connect url information or retrieve them from server

test_in_shinyappsio: Test connect applications given the server and account

Examples

if (FALSE) test_in_connect(dir = "apps")
if (FALSE) test_in_test_in_shinyapps_io(dir = "apps")