Test Apps in SSO/SSP

Automatically launches docker in a background process. Once the docker is ready, a shiny application will be launched to help move through the applications.

test_in_sso(
  dir = "apps",
  apps = apps_manual(dir),
  app = apps[1],
  release = c("focal", "bionic", "xenial", "centos7"),
  r_version = c("4.1", "4.0", "3.6", "3.5"),
  tag = NULL,
  port = 8080,
  port_background = switch(release, centos7 = 7878, 3838),
  host = "127.0.0.1"
)

test_in_ssp(
  dir = "apps",
  apps = apps_manual(dir),
  app = apps[1],
  release = c("focal", "bionic", "xenial", "centos7"),
  r_version = c("4.1", "4.0", "3.6", "3.5"),
  tag = NULL,
  port = 8080,
  port_background = switch(release, centos7 = 8989, 4949),
  host = "127.0.0.1"
)

Arguments

dir
base folder to look for applications

apps
applications within dir to run

app
app number or name to start with. If numeric, it will match the leading number in the testing application

release
Distro release name, such as "focal" for ubuntu or "7" for centos

r_version
R version to use. Ex: "3.6"

tag
Extra tag information for the docker image. This will prepend a - if a value is given.

port
Port for shiny application

port_background
Port to connect to the Docker container

host
The IPv4 address that the application should listen on. Defaults to the shiny.host option, if set, or "127.0.0.1" if not. See Details.

Details

The docker application will stop when the shiny application exits.

Functions

test_in_sso: Test SSO Shiny applications

test_in_ssp: Test SSP Shiny applications

Examples

if (FALSE) test_in_connect(dir = "apps")