Test deployed apps

Automatically runs the next app in a fresh callr::r_bg session. To stop, close the shiny application window.

test_in_connect(
  app_name = apps[1],
  apps = apps_manual,
  ...,
  urls = default_connect_urls,
  server = attr(urls, "server"),
  account = attr(urls, "account"),
  port = 8080,
  host = "127.0.0.1"
)

test_in_shinyappsio(
  app_name = apps[1],
  apps = apps_manual,
  port = 8080,
  host = "127.0.0.1",
  account = "testing-apps",
  server = "shinyapps.io"
)

Arguments

app_name

app number or name to start with. If numeric, it will match the leading number in the testing application

apps

List of apps to test

...

ignored

urls

Named vector of urls to visit. This should be the output of [connect_urls]. By default this will set server, account, dir, and apps

server, account

Parameters that could be supplied to [rsconnect::deployApp]

port

port for the foreground app process

host

host for the foreground app processes

Functions

test_in_connect: Test connect applications given the server and account

test_in_shinyappsio: Test connect applications given the server and account

Examples

if (FALSE) test_in_connect(dir = "apps")
if (FALSE) test_in_test_in_shinyapps_io(dir = "apps")